app/requirements.prod.txt
Django==5.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
django-environ==0.11.2
django-import-export==4.3.7
openpyxl==3.1.5
django-simple-history==3.8.0
drf-spectacular==0.28.0
django-unfold==0.55.3
django-cachalot==2.7.0
Pillow==10.4.0
psycopg2-binary~=2.9.10
gunicorn==23.0.0
django-redis==5.4.0
